requirements.txt
google-generativeai==0.2.0
google-ai-generativelanguage==0.3.2
tqdm==4.66.1
google-api-core[grpc]==2.11.1
proto-plus==1.22.3
protobuf==3.20.3
googleapis-common-protos==1.60.0
google-auth==2.17.3
requests==2.31.0
grpcio==1.57.0
grpcio-status==1.48.2
cachetools==5.3.1
pyasn1-modules>=0.2.1
six>=1.9.0
rsa<5,>=3.1.4
charset-normalizer<4,>=2
idna<4,>=2.5
urllib3<3,>=1.21.1
certifi>=2017.4.17
pyasn1<0.6.0,>=0.5.0
streamlit==1.12.0
